refactor(db, parser, ci): post-review improvements (#6)
* refactor(db, parser, ci): post-review improvements

- Replace flogging with logging.New; ban flogging via depguard
- Fix DSN via dbconn.DataSourceName; fix rollback to use context.Background()
- Remove ProcessedBlock.Txns and TxRecord.BlockNum; derive at write time
- Move buildBatchParams before beginTx; add nil guards for BlockInfo and Data
- Pre-allocate all seven batchParams slices
- Change TxRecord.ValidationCode from int32 to protoblocktx.Status (typed)
- Change validation_code column from BIGINT to SMALLINT
- Log warning when endorsement identity parsing fails (was silently swallowed)
- Add util.Ptr[T] generic helper; remove scattered ptr helpers
- Rename test_exports.go to test_exports_test.go
- Add TestMain with DB_DEPLOYMENT=local and DB_TYPE=postgres
- CI: add lint and test jobs; pin all action SHAs; add permissions block;
  set timeout-minutes; pin golangci-lint to v2.10.1; verify sqlc SHA256
- Makefile: add start-db/stop-db targets


* ci: upgrade golangci-lint-action v6→v9 for golangci-lint v2 support


* ci: fix golangci-lint v2 config schema violations

Move top-level 'exclusions' under 'linters' and top-level 'sort-order'
under 'output', matching the v2 JSON schema.


---------

+++ pkg/db/schema.sql
@@ -9,7 +9,7 @@ CREATE TABLE IF NOT EXISTS transactions (
     block_num BIGINT NOT NULL REFERENCES blocks(block_num),
     tx_num    BIGINT NOT NULL,
     tx_id     BYTEA  NOT NULL,
-    validation_code BIGINT NOT NULL,
+    validation_code SMALLINT NOT NULL,
     PRIMARY KEY (block_num, tx_num)
 );
