+++ pkg/db/sqlc/batch.go
@@ -203,7 +203,7 @@ type InsertTransactionParams struct {
 	BlockNum       int64  `json:"block_num"`
 	TxNum          int64  `json:"tx_num"`
 	TxID           []byte `json:"tx_id"`
-	ValidationCode int64  `json:"validation_code"`
+	ValidationCode int16  `json:"validation_code"`
 }
 
 func (q *Queries) InsertTransaction(ctx context.Context, arg []InsertTransactionParams) *InsertTransactionBatchResults {
